move: main or post-form
