+++ src/app/(main)/welcome/page.tsx
